Prototype Project Demo/Assessment Project/requirements.txt
asgiref==3.7.2
celery==5.4.0
channels==4.1.0
Django==5.0.2
djangorestframework==3.15.1
injector==0.21.0
Pygments==2.17.2
pytz==2024.1
Requests==2.31.0
setuptools==69.5.1
